+++ tests/sdk/test_download.py
@@ -31,7 +31,6 @@ from ..utils import mock_response
     [
         ("dataset", "opener.py"),
         ("algo", "algo.tar.gz"),
-        ("metric", "metrics.py"),
         ("model", "model_foo"),
     ],
 )
@@ -51,7 +50,7 @@ def test_download_asset(asset_type, filename, tmp_path, client, mocker):
     m.assert_called()
 
 
-@pytest.mark.parametrize("asset_type", ["dataset", "algo", "metric", "model", "logs"])
+@pytest.mark.parametrize("asset_type", ["dataset", "algo", "model", "logs"])
 def test_download_asset_not_found(asset_type, tmp_path, client, mocker):
     m = mock_requests(mocker, "get", status=404)
 
@@ -62,7 +61,7 @@ def test_download_asset_not_found(asset_type, tmp_path, client, mocker):
     assert m.call_count == 1
 
 
-@pytest.mark.parametrize("asset_type", ["dataset", "algo", "metric", "model"])
+@pytest.mark.parametrize("asset_type", ["dataset", "algo", "model"])
 def test_download_content_not_found(asset_type, tmp_path, client, mocker):
     item = getattr(datastore, asset_type.upper())